# static/swagger.yaml
openapi: 3.0.3
info:
  title: md-to-pdf Document Engine API
  description: |
    Moteur de documents complet : conversion Markdown/HTML/Template vers PDF,
    thèmes, graphiques et diagrammes, aperçu PNG, fusion, filigrane, protection
    par mot de passe, caviardage, audit de mise en page et comparaison visuelle.

    ## Comportement conditionnel des réponses

    La plupart des endpoints de génération PDF acceptent les champs optionnels
    `client_id` et `pdf_name`. Selon leur présence :

    | `client_id` + `pdf_name` | Réponse |
    |--------------------------|---------|
    | **fournis** | Le PDF est sauvegardé sur le serveur et un JSON `{ "download_url": "/download/…" }` est retourné |
    | **absents** | Le fichier PDF binaire est retourné directement dans le body |

    `client_id` et `pdf_name` servent à construire un chemin sur disque : ils doivent être
    des noms simples (`[A-Za-z0-9._-]`, 128 caractères max, ne commençant pas par un point).

    ## Deux façons de désigner un PDF

    Partout où un corps de requête attend un PDF — `pdf`, `before`, `after`, les entrées de
    `pdfs` — **deux formes sont acceptées** :

    | Référence | Origine |
    |-----------|---------|
    | `/download/{client_id}/{nom}.pdf` | un PDF que ce service a produit et sauvegardé |
    | `asset://as_…` | un fichier déposé par l'appelant via `POST /api/files` |

    La première n'a pas changé et ne changera pas. La seconde est ce qui rend les outils
...

    ## Sortie des outils

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyHeader: []
  - BearerToken: []
...
paths:
...
  /api/preview:
    post:
      tags: [Preview]
      summary: Génération de preview PNG
      description: |
        Génère un PDF à partir de l'un des trois modes d'entrée, puis en retourne
        une ou plusieurs pages en PNG.

        **Modes d'entrée** (mutuellement exclusifs, par ordre de priorité) :
        1. `markdown` — conversion Markdown → PDF → PNG
        2. `template` + `data` — rendu Tera → PDF → PNG
        3. `html` — HTML → PDF → PNG

        **Rétro-compatibilité** : un corps qui ne porte aucun des champs `pages`, `dpi`
        ou `layout` retourne toujours l'image PNG de la première page, comme avant.

        Sans `layout` explicite, une seule page donne un `image/png` brut et plusieurs
        pages donnent le JSON `images`.
      operationId: preview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewRequest"
            examples:
              markdown:
                summary: Preview d'un document Markdown
                value:
                  markdown: "# Aperçu\n\nTexte de test."
              html:
                summary: Preview d'un HTML
                value:
                  html: "<html><body><h1>Preview HTML</h1></body></html>"
              template:
                summary: Preview d'un template Tera
                value:
                  template: "<h1>{{ title }}</h1>"
                  data:
                    title: "Mon titre"
      responses:
        "200":
          description: |
            Image PNG (page unique ou planche contact), ou JSON `images` pour
            plusieurs pages.
          content:
            image/png:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewPages"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Merge
  # ───────────────────────────────────────────────
...
  /api/watermark:
    post:
      tags: [Watermark]
      summary: Ajouter un watermark texte
      description: |
        Ajoute un watermark texte semi-transparent en overlay sur un PDF existant.

        Le PDF source est un chemin `/download/…` **ou** une référence `asset://as_…`.
      operationId: watermark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatermarkRequest"
            example:
              pdf: "/download/client-123/rapport.pdf"
              text: "BROUILLON"
              opacity: 0.1
              angle: -45
              client_id: "client-123"
              pdf_name: "rapport-watermarked"
      responses:
        "200":
          description: PDF avec watermark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertResponse"
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Protect
  # ───────────────────────────────────────────────
...
  /api/compress:
    post:
      tags: [Compress]
      summary: Compresser un PDF, et dire ce que ça a coûté
      description: |
        Ghostscript en `-dSAFER`, quatre niveaux, et la facture attachée : le `verdict`
        compare le nombre de pages et le nombre de caractères extractibles avant et après.
        Un scan poussé en `screen` revient quatre fois plus léger et n'est plus cherchable ;
        c'est exactement ce que le verdict nomme.

        Une compression qui produit un fichier **plus gros** est une compression ratée : la
        source est alors renvoyée telle quelle, et le verdict le dit.
      operationId: compress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompressRequest"
            example:
              pdf: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
              level: ebook
              output: asset
      responses:
        "200":
          $ref: "#/components/responses/ToolResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Repair
  # ───────────────────────────────────────────────
...
  /download/{client_id}/{pdf_name}:
    get:
      tags: [Download]
      summary: Télécharger un PDF sauvegardé
      description: |
        Télécharge un fichier PDF précédemment sauvegardé sur le serveur.
        Le header `Content-Disposition` force le téléchargement.
      operationId: downloadPdf
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Identifiant du client
          example: "client-123"
        - name: pdf_name
          in: path
          required: true
          schema:
            type: string
          description: Nom du fichier PDF
          example: "rapport-2024.pdf"
      responses:
        "200":
          description: Fichier PDF
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="rapport-2024.pdf"'
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Fichier non trouvé

  # ───────────────────────────────────────────────
  # Legacy  (POST / FormData)
  # ───────────────────────────────────────────────
...
components:
...
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Clé d'API, active uniquement si `API_KEY` est définie côté serveur.
    BearerToken:
      type: http
      scheme: bearer
      description: |
        Même clé que `X-API-Key`, transmise via `Authorization: Bearer <clé>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: mdpdf_session
      description: |
        Cookie de session posé par `POST /api/auth/signup` et `POST /api/auth/login`.
        `HttpOnly`, `SameSite=Lax`, et `Secure` dès que `X-Forwarded-Proto` vaut `https`.

        Il ne se mélange jamais avec `X-API-Key` : le cookie dit « un humain regarde une
        page », l'en-tête dit « un programme appelle ». C'est cette séparation qui permet au
        cookie de rester `Lax` — donc de refuser de voyager sur un POST cross-site — pendant
        que la clé reste utilisable depuis n'importe où.

  # ─── Paramètres réutilisables ─────────────────
...
  schemas:
    PdfSource:
      type: string
      description: |
        Référence d'un PDF, sous **l'une des deux formes acceptées** :

        - `/download/{client_id}/{nom}.pdf` — un PDF que ce service a produit et sauvegardé ;
        - `asset://as_…` — un fichier déposé via `POST /api/files`.

        La première forme est celle d'hier, inchangée. La seconde est ce qui rend l'endpoint
        utilisable sur un document que le service n'a pas fabriqué.
      example: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
...
    PdfEngine:
      type: string
      enum: [weasyprint, wkhtmltopdf, pdflatex]
      default: weasyprint
      description: Moteur PDF utilisé par pandoc
    PaperSize:
      type: string
      enum: [a4, a3, letter]
      default: a4
    Orientation:
      type: string
      enum: [portrait, landscape]
      default: portrait
...
    PdfOptions:
      type: object
      description: Options de rendu PDF communes à tous les endpoints de génération
      properties:
        paper_size:
          $ref: "#/components/schemas/PaperSize"
        orientation:
          $ref: "#/components/schemas/Orientation"
        margins:
          $ref: "#/components/schemas/Margins"
        page_numbers:
          type: boolean
          default: false
          description: Afficher les numéros de page en bas de chaque page
        page_number_format:
          type: string
          default: "counter(page)"
          description: Format CSS du numéro de page (ex. `counter(page) "/" counter(pages)`)
        toc:
          type: boolean
          default: false
          description: Générer une table des matières (pandoc uniquement)
        toc_depth:
          type: integer
          minimum: 1
          maximum: 6
          description: Profondeur de la table des matières
        watermark:
          type: string
          description: Texte de watermark léger appliqué via CSS
        theme:
          type: string
          description: |
            Kit de marque : `"nom"` pour la dernière version, `"nom@2"` pour figer.
            Voir `GET /api/themes`. Un thème publié est immuable.
          example: report
        cover:
          $ref: "#/components/schemas/Cover"
        autolayout:
          type: boolean
          default: false
          description: |
            Analyse le PDF produit et le re-rend avec du CSS correctif tant que le score
            progresse. Le rapport est renvoyé dans le champ `layout` de la réponse.
        censor_label:
          type: string
          description: |
            Libellé des blocs CENSOR qui ne nomment pas de niveau. Cascade :
            niveau du tag > `censor_label` > libellé historique.
        charts:
          type: boolean
          default: true
          description: "`false` coupe l'expansion des blocs ```chart ET ```mermaid"
...
    PreviewRequest:
      type: object
      description: |
        Fournir **un seul** des modes d'entrée :
        - `markdown` pour une conversion Markdown
        - `template` + `data` pour un rendu Tera
        - `html` pour du HTML direct
      properties:
        markdown:
          type: string
          description: Contenu Markdown (mode Markdown)
        html:
          type: string
          description: Contenu HTML (mode HTML)
        template:
          type: string
          description: Template Tera (mode Template, requiert `data`)
        data:
          type: object
          additionalProperties: true
          description: Données JSON pour le template (mode Template)
        css:
          type: string
        engine:
          $ref: "#/components/schemas/PdfEngine"
        options:
          $ref: "#/components/schemas/PdfOptions"
        pages:
          type: string
          description: '`"3"`, `"2-5"` ou `"all"`. Absent : la page 1 seule.'
          example: "1-4"
        dpi:
          type: integer
          minimum: 36
          maximum: 300
          default: 150
        layout:
          type: string
          enum: [png, images, sheet]
          description: |
            `png` image brute, `images` un PNG par page en JSON, `sheet` une planche
            contact en une seule image. Absent : `png` pour une page, `images` au-delà.
        header_html:
          type: string
        footer_html:
          type: string
        header_template:
          type: string
        footer_template:
          type: string
...
    WatermarkRequest:
      type: object
      required: [pdf, text]
      properties:
        pdf:
          $ref: "#/components/schemas/PdfSource"
        text:
          type: string
          description: Texte du watermark
        opacity:
          type: number
          format: float
          default: 0.06
          minimum: 0
          maximum: 1
          description: Opacité du watermark (0 = invisible, 1 = opaque)
        angle:
          type: number
          format: float
          default: -45
          description: Angle de rotation du watermark en degrés
        client_id:
          type: string
        pdf_name:
          type: string
...
    PreviewPages:
      type: object
      properties:
        pages:
          type: array
          items:
            type: object
            properties:
              page:
                type: integer
                description: Numéro de page 1-indexé dans le document source
              png:
                type: string
                description: Image encodée en base64
              width:
                type: integer
              height:
                type: integer
        pages_total:
          type: integer
          description: Pages du document, qui peuvent être plus nombreuses que celles rendues
        truncated:
          type: boolean
          description: Présent uniquement quand le plafond de pages a écourté la réponse
    ConvertResponse:
      type: object
      description: |
        Strictement additive : une requête qui ne demande rien de nouveau reçoit
        exactement `{ "download_url": "…" }`, comme avant.
      properties:
        download_url:
          type: string
          description: URL relative pour télécharger le PDF généré
          example: "/download/client-123/rapport-2024.pdf"
        cached:
          type: boolean
          description: Présent (et vrai) uniquement quand le PDF vient du cache de rendu
        layout:
          allOf:
            - $ref: "#/components/schemas/LayoutReport"
          description: Présent uniquement quand `options.autolayout` a été demandé
        warnings:
          type: array
          description: |
            Présent uniquement s'il y en a : un bloc `chart` ou `mermaid` qui n'a pas pu
            être rendu est laissé tel quel dans le document et la raison arrive ici.
          items:
            $ref: "#/components/schemas/BlockWarning"
    BlockWarning:
      type: object
      properties:
        kind:
          type: string
          enum: [chart, mermaid]
        message:
          type: string
        line:
          type: integer
          description: Ligne 1-indexée du bloc fautif dans le document source
    LayoutReport:
      type: object
      properties:
        pages:
          type: integer
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: 100 = aucun défaut détecté
        passes:
          type: integer
          description: Nombre de passes correctives appliquées (autolayout uniquement)
        issues:
          type: array
          items:
            type: object
            properties:
              kind:
                type: string
                enum: [overflow, blank_page, widow_page, orphan_heading, split_table, long_table]
              page:
                type: integer
                description: Numéro de page 1-indexé
              severity:
                type: string
                enum: [error, warn, info]
              detail:
                type: string
              bbox:
                type: array
                description: x0, y0, x1, y1 en points PDF, quand le défaut est localisé
                items:
                  type: number

    # ─── Assets et travaux ──────────────────────
...
    CompressRequest:
      type: object
      required: [pdf]
      properties:
        pdf:
          $ref: "#/components/schemas/PdfSource"
        level:
          type: string
          enum: [screen, ebook, printer, prepress]
          default: ebook
          description: |
            Préréglage Ghostscript. Refuser une valeur inconnue est délibéré :
            `-dPDFSETTINGS` accepte n'importe quelle chaîne et ignore silencieusement celles
            qu'il ne connaît pas, donc une faute de frappe reviendrait en fichier jamais
            compressé.
        dpi:
          type: integer
          minimum: 36
          maximum: 1200
          description: |
            Résolution de rééchantillonnage des images. Absent, celle du préréglage
            s'applique.
        client_id:
          type: string
        pdf_name:
          type: string
        output:
          $ref: "#/components/schemas/ToolOutput"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Type d'erreur
          example: "Bad request"
        details:
          type: string
          description: Détails / message d'erreur
          example: "At least 2 PDFs are required for merging"
...
  responses:
    ToolResult:
      description: |
        Le fichier produit, ou le JSON `ToolResponse` quand `client_id` + `pdf_name` ou
        `"output":"asset"` ont été demandés.
      content:
        application/pdf:
          schema:
            type: string
            format: binary
        application/json:
          schema:
            $ref: "#/components/schemas/ToolResponse"
...
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Bad request"
            details: "\"data\" must be a JSON object"
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Not found"
            details: "PDF not found: /download/client-123/missing.pdf"
    Unauthorized:
      description: Clé d'API manquante ou invalide (uniquement si `API_KEY` est configurée)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Unauthorized"
            details: "Missing or invalid API key (X-API-Key or Authorization: Bearer)"
    TooManyRequests:
      description: |
        La file de rendu est pleine : la requête a attendu plus de
        `PDF_QUEUE_TIMEOUT_SECS` sans obtenir de place. Réessayer quelques secondes plus tard.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Timeout:
      description: La génération a dépassé la limite de temps (`PDF_PROCESS_TIMEOUT_SECS`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Timeout"
            details: "weasyprint exceeded the 60s time limit"
    InternalError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Pandoc conversion failed"
            details: "pandoc: ... (stderr output)"